moved infographic to direct hazards
